Reorganize clue files
