+++ src/clues/__tests__/start-with.test.ts
@@ -1,4 +1,4 @@
-import { checkClue } from "../../../clues.js";
+import { checkClue } from "../../clues.js";
 
 test("startWith", () => {
   expect(checkClue("spacer image.")).toMatchInlineSnapshot(`
