+++ src/clues/end-punctuation.ts
@@ -1,5 +1,5 @@
-import { Rule, createSuggestion } from "../../clues.js";
-import { Alt } from "../../index.js";
+import { Rule, createSuggestion } from "../clues.js";
+import { Alt } from "../index.js";
 
 export function checkPunctuation(alt: Alt) {
   return !/[.!?]$/.test(alt) && alt.length > 1
