+++ src/clues/end-with.ts
@@ -1,4 +1,4 @@
-import { Rule } from "../../clues.js";
+import { Rule } from "../clues.js";
 
 export default function atlEndsWith(): Rule {
   return {
